+++ src/app.module.ts
@@ -4,12 +4,13 @@ import { AppService } from './app.service';
 import { ReservationsModule } from './modules/reservations/reservations.module';
 import { ConfigModule } from '@nestjs/config';
 import { DatesModule } from './modules/dates/dates.module';
-import { GoogleSheetsModule } from './google-sheets/google-sheets.module';
+import { GoogleSheetsModule } from './modules/google-sheets/google-sheets.module';
 import { AiModule } from './modules/ai/ai.module';
+import { WhatsAppModule } from './modules/whats-app/whats-app.module';
 
 
 @Module({
-  imports: [ReservationsModule, GoogleSheetsModule.forRoot(), ConfigModule.forRoot({ envFilePath: '.env' }), DatesModule, AiModule],
+  imports: [ReservationsModule, GoogleSheetsModule.forRoot(), ConfigModule.forRoot({ envFilePath: '.env' }), DatesModule, AiModule, WhatsAppModule],
   controllers: [AppController],
   providers: [AppService],
 })