+++ src/modules/dates/application/delete-reservation.use-case.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@nestjs/common";
 import { SHEETS_NAMES } from "src/constants";
-import { GoogleSheetsService } from "src/google-sheets/service/google-sheets.service";
+import { GoogleSheetsService } from "src/modules/google-sheets/service/google-sheets.service";
 import { DeleteReservation, GetIndexParams, ReservationOperation, UpdateParams } from "src/lib"
 import { Logger } from "@nestjs/common";
 import { GenerateDatetime } from "../dateTime-build/generate-datetime";