+++ src/modules/dates/dates.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { DatesService } from './service/dates.service';
 import { DatesController } from './controller/dates.controller';
-import { GoogleSheetsModule } from 'src/google-sheets/google-sheets.module';
+import { GoogleSheetsModule } from 'src/modules/google-sheets/google-sheets.module';
 import { GenerateDatetime } from './dateTime-build/generate-datetime';
 import { CreateDayUseCase, CreateReservationRowUseCase, DeleteReservationUseCase } from './application';