+++ src/modules/whats-app/controller/whats-app.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Post, Body, Get, Query, Delete } from '@nestjs/common';
+import { WhatsAppService } from '../service/whats-app.service';
+
+@Controller('communication')
+export class WhatsAppController {
+  constructor(private readonly whatsappService: WhatsAppService) {}
+
+  @Post()
+  create(): Promise<string> {
+    return this.whatsappService.createReservation();
+  }
+}